refactor view methods to DataAssistantResult
