+++ great_expectations/rule_based_profiler/data_assistant_result/data_assistant_result.py
@@ -4,6 +4,7 @@ import copy
 import datetime
 import json
 import os
+import pprint
 import uuid
 from collections import defaultdict
 from dataclasses import asdict, dataclass, field
@@ -18,6 +19,7 @@ from typing import (
     NamedTuple,
     Optional,
     Set,
+    Tuple,
     Union,
 )
 
@@ -75,6 +77,7 @@ from great_expectations.types import (
     SerializableDictDot,
     TintsAndShades,
 )
+from great_expectations.util import deep_filter_properties_iterable
 
 if TYPE_CHECKING:
     from great_expectations.core import (
@@ -4284,3 +4287,281 @@ Use DataAssistantResult.metrics_by_domain to show all calculated Metrics"""
             sanitized_metric_names=sanitized_metric_names,
             sequential=sequential,
         )
+
+
+class ExpectationSuiteViewer:
+    """Legacy methods to print Expectations within an ExpectationSuite"""
+
+    def __init__(self, suite: ExpectationSuite):
+        self._suite = suite
+
+    @public_api
+    def show_expectations_by_domain_type(self) -> None:
+        """Displays "ExpectationConfiguration" list, grouped by "domain_type", in predetermined designated order.
+
+        The means of displaying is through the use of the "Pretty Print" library method "pprint.pprint()".
+        """
+        expectation_configurations_by_domain: Dict[
+            str, List[ExpectationConfiguration]
+        ] = self.get_grouped_and_ordered_expectations_by_domain_type()
+
+        domain_type: str
+        expectation_configurations: List[ExpectationConfiguration]
+        for (
+            domain_type,
+            expectation_configurations,
+        ) in expectation_configurations_by_domain.items():
+            pprint.pprint(object=MetricDomainTypes(domain_type).value.capitalize())
+            self.show_expectations_by_expectation_type(
+                expectation_configurations=expectation_configurations
+            )
+
+    def show_expectations_by_expectation_type(
+        self,
+        expectation_configurations: Optional[List[ExpectationConfiguration]] = None,
+    ) -> None:
+        """Displays "ExpectationConfiguration" list, grouped by "expectation_type", in predetermined designated order.
+
+        The means of displaying is through the use of the "Pretty Print" library method "pprint.pprint()".
+        """
+        if expectation_configurations is None:
+            expectation_configurations = (
+                self.get_grouped_and_ordered_expectations_by_expectation_type()
+            )
+
+        expectation_configuration: ExpectationConfiguration
+        domain_type: MetricDomainTypes
+        kwargs: dict
+        pprint_objects: List[dict] = []
+        for expectation_configuration in expectation_configurations:
+            domain_type = expectation_configuration.get_domain_type()
+            kwargs = expectation_configuration.kwargs
+            pprint_objects.append(
+                {
+                    expectation_configuration.expectation_type: {
+                        "domain": domain_type.value,
+                        **kwargs,
+                    }
+                }
+            )
+        pprint.pprint(
+            object=pprint_objects,
+            indent=2,
+        )
+
+    def get_grouped_and_ordered_expectations_by_domain_type(
+        self,
+    ) -> Dict[str, List[ExpectationConfiguration]]:
+        """
+        Returns "ExpectationConfiguration" list in predetermined order by passing appropriate methods for retrieving
+        "ExpectationConfiguration" lists by corresponding "domain_type" (with "table" first; then "column", and so on).
+        """
+        expectation_configurations_by_domain: Dict[
+            str, List[ExpectationConfiguration]
+        ] = self._get_expectations_by_domain_using_accessor_method(
+            domain_type=MetricDomainTypes.TABLE.value,
+            accessor_method=self.get_table_expectations,
+        )
+        expectation_configurations_by_domain.update(
+            self._get_expectations_by_domain_using_accessor_method(
+                domain_type=MetricDomainTypes.COLUMN.value,
+                accessor_method=self.get_column_expectations,
+            )
+        )
+        expectation_configurations_by_domain.update(
+            self._get_expectations_by_domain_using_accessor_method(
+                domain_type=MetricDomainTypes.COLUMN_PAIR.value,
+                accessor_method=self.get_column_pair_expectations,
+            )
+        )
+        expectation_configurations_by_domain.update(
+            self._get_expectations_by_domain_using_accessor_method(
+                domain_type=MetricDomainTypes.MULTICOLUMN.value,
+                accessor_method=self.get_multicolumn_expectations,
+            )
+        )
+        return expectation_configurations_by_domain
+
+    def get_grouped_and_ordered_expectations_by_expectation_type(
+        self,
+    ) -> List[ExpectationConfiguration]:
+        """
+        Returns "ExpectationConfiguration" list, grouped by "expectation_type", in predetermined designated order.
+        """
+        table_expectation_configurations: List[ExpectationConfiguration] = sorted(
+            self.get_table_expectations(),
+            key=lambda element: element["expectation_type"],
+        )
+        column_expectation_configurations: List[ExpectationConfiguration] = sorted(
+            self.get_column_expectations(),
+            key=lambda element: element["expectation_type"],
+        )
+        column_pair_expectation_configurations: List[ExpectationConfiguration] = sorted(
+            self.get_column_pair_expectations(),
+            key=lambda element: element["expectation_type"],
+        )
+        multicolumn_expectation_configurations: List[ExpectationConfiguration] = sorted(
+            self.get_multicolumn_expectations(),
+            key=lambda element: element["expectation_type"],
+        )
+        return (
+            table_expectation_configurations
+            + column_expectation_configurations
+            + column_pair_expectation_configurations
+            + multicolumn_expectation_configurations
+        )
+
+    def get_table_expectations(self) -> List[ExpectationConfiguration]:
+        """Return a list of table expectations."""
+        expectation_configurations: List[ExpectationConfiguration] = list(
+            filter(
+                lambda element: element.get_domain_type() == MetricDomainTypes.TABLE,
+                self._suite.expectation_configurations,
+            )
+        )
+
+        expectation_configuration: ExpectationConfiguration
+        for expectation_configuration in expectation_configurations:
+            expectation_configuration.kwargs = deep_filter_properties_iterable(
+                properties=expectation_configuration.kwargs, clean_falsy=True
+            )
+
+        return expectation_configurations
+
+    def get_column_expectations(self) -> List[ExpectationConfiguration]:
+        """Return a list of column map expectations."""
+        expectation_configurations: List[ExpectationConfiguration] = list(
+            filter(
+                lambda element: element.get_domain_type() == MetricDomainTypes.COLUMN,
+                self._suite.expectation_configurations,
+            )
+        )
+
+        expectation_configuration: ExpectationConfiguration
+        kwargs: dict
+        column_name: str
+        for expectation_configuration in expectation_configurations:
+            kwargs = deep_filter_properties_iterable(
+                properties=expectation_configuration.kwargs, clean_falsy=True
+            )
+            column_name = kwargs.pop("column")
+            expectation_configuration.kwargs = {"column": column_name, **kwargs}
+
+        return expectation_configurations
+
+    # noinspection PyPep8Naming
+    def get_column_pair_expectations(self) -> List[ExpectationConfiguration]:
+        """Return a list of column_pair map expectations."""
+        expectation_configurations: List[ExpectationConfiguration] = list(
+            filter(
+                lambda element: element.get_domain_type()
+                == MetricDomainTypes.COLUMN_PAIR,
+                self._suite.expectation_configurations,
+            )
+        )
+
+        expectation_configuration: ExpectationConfiguration
+        kwargs: dict
+        column_A_name: str
+        column_B_name: str
+        for expectation_configuration in expectation_configurations:
+            kwargs = deep_filter_properties_iterable(
+                properties=expectation_configuration.kwargs, clean_falsy=True
+            )
+            column_A_name = kwargs.pop("column_A")
+            column_B_name = kwargs.pop("column_B")
+            expectation_configuration.kwargs = {
+                "column_A": column_A_name,
+                "column_B": column_B_name,
+                **kwargs,
+            }
+
+        return expectation_configurations
+
+    def get_multicolumn_expectations(self) -> List[ExpectationConfiguration]:
+        """Return a list of multicolumn map expectations."""
+        expectation_configurations: List[ExpectationConfiguration] = list(
+            filter(
+                lambda element: element.get_domain_type()
+                == MetricDomainTypes.MULTICOLUMN,
+                self._suite.expectation_configurations,
+            )
+        )
+
+        expectation_configuration: ExpectationConfiguration
+        kwargs: dict
+        column_list: str
+        for expectation_configuration in expectation_configurations:
+            kwargs = deep_filter_properties_iterable(
+                properties=expectation_configuration.kwargs, clean_falsy=True
+            )
+            column_list = kwargs.pop("column_list")
+            expectation_configuration.kwargs = {"column_list": column_list, **kwargs}
+
+        return expectation_configurations
+
+    def get_grouped_and_ordered_expectations_by_column(
+        self, expectation_type_filter: Optional[str] = None
+    ) -> Tuple[Dict[str, List[ExpectationConfiguration]], List[str]]:
+        expectations_by_column: Dict[str, List[ExpectationConfiguration]] = {}
+        ordered_columns: List[str] = []
+
+        column: str
+        expectation: ExpectationConfiguration
+        for expectation in self._suite.expectation_configurations:
+            if "column" in expectation.kwargs:
+                column = expectation.kwargs["column"]
+            else:
+                column = "_nocolumn"
+
+            if column not in expectations_by_column:
+                expectations_by_column[column] = []
+
+            if (
+                expectation_type_filter is None
+                or expectation.expectation_type == expectation_type_filter
+            ):
+                expectations_by_column[column].append(expectation)
+
+            # if possible, get the order of columns from expect_table_columns_to_match_ordered_list
+            if (
+                expectation.expectation_type
+                == "expect_table_columns_to_match_ordered_list"
+            ):
+                exp_column_list: List[str] = expectation.kwargs["column_list"]
+                if exp_column_list and len(exp_column_list) > 0:
+                    ordered_columns = exp_column_list
+
+        # Group items by column
+        sorted_columns = sorted(list(expectations_by_column.keys()))
+
+        # only return ordered columns from expect_table_columns_to_match_ordered_list evr if they match set of column
+        # names from entire evr, else use alphabetic sort
+        if set(sorted_columns) == set(ordered_columns):
+            return expectations_by_column, ordered_columns
+
+        return expectations_by_column, sorted_columns
+
+    @staticmethod
+    def _get_expectations_by_domain_using_accessor_method(
+        domain_type: str, accessor_method: Callable
+    ) -> Dict[str, List[ExpectationConfiguration]]:
+        expectation_configurations_by_domain: Dict[
+            str, List[ExpectationConfiguration]
+        ] = {}
+
+        expectation_configurations: List[ExpectationConfiguration]
+        expectation_configuration: ExpectationConfiguration
+        for expectation_configuration in accessor_method():
+            expectation_configurations = expectation_configurations_by_domain.get(  # type: ignore[assignment]
+                domain_type
+            )
+            if expectation_configurations is None:
+                expectation_configurations = []
+                expectation_configurations_by_domain[
+                    domain_type
+                ] = expectation_configurations
+
+            expectation_configurations.append(expectation_configuration)
+
+        return expectation_configurations_by_domain
